Fix Filament filters to avoid contradictory query clauses; guard SelectFilter closures in BonEntrees & BonSorties tables; minor form numeric casts; update copilot instructions

+++ app/Filament/Resources/BonEntrees/Pages/CreateBonEntree.php
@@ -26,8 +26,8 @@ class CreateBonEntree extends CreateRecord
         $bobines = $data['bobineItems'] ?? [];
         $products = $data['productItems'] ?? [];
 
-        $bobinesHt = collect($bobines)->sum(fn ($item) => $item['price_ht'] ?? 0);
-        $productsHt = collect($products)->sum(fn ($item) => ($item['qty_entered'] ?? 0) * ($item['price_ht'] ?? 0));
+    $bobinesHt = collect($bobines)->sum(fn ($item) => (float) ($item['price_ht'] ?? 0));
+    $productsHt = collect($products)->sum(fn ($item) => (float) ($item['qty_entered'] ?? 0) * (float) ($item['price_ht'] ?? 0));
 
         $data['total_amount_ht'] = $bobinesHt + $productsHt;
         $data['total_amount_ttc'] = $data['total_amount_ht'] + ($data['frais_approche'] ?? 0);
